Move around setup script

+++ setup.sh
@@ -7,26 +7,15 @@ sudo apt-get install vim -y
 sudo apt-get install curl -y
 sudo apt-get install cmake -y
 
-# Install Oh-My-Zsh
-sudo apt-get install zsh -y
-sh -c "$(curl -fsSL https://raw.githubusercontent.com/robbyrussell/oh-my-zsh/master/tools/install.sh)"
-sudo chsh -s $(which zsh)
-sudo apt-get install fonts-powerline -y
-
 # More tools
 sudo apt-get install hdf5-tools -y
 sudo apt-get install hdfview -y
 sudo apt-get install ffmpeg -y
 sudo apt-get install libeigen3-dev -y
-sudo apt-get install gnome-tweak-tools -y
+sudo apt-get install gnome-tweaks -y
 
-echo "Install Python (y/n)?"
-read check_install
-if [ $check_install = "y" ]; then
-    sudo apt install python3-pip -y
-    pip3 install --upgrade pip
-    pip3 install jupyter
-fi
+# Install Vundle
+git clone https://github.com/VundleVim/Vundle.vim.git ~/.vim/bundle/Vundle.vim
 
 echo "Install Jekyll (y/n)?"
 read check_install
@@ -39,25 +28,6 @@ if [ $check_install = "y" ]; then
     gem install jekyll bundler
 fi
 
-echo "Install Spotify (y/n)?"
-read spotify_install
-if [ $spotify_install = "y" ]; then
-    sudo apt install snapd
-   snap install spotify
-fi
-
-echo "Install Atom (y/n)?"
-read atom_install
-if [ $atom_install = "y"]; then
-    snap install atom --classic
-fi
-
-echo "Install Chromium (y/n)?"
-read check_install
-if [ $check_install = "y" ]; then
-    sudo apt-get install chromium-browser -y
-fi
-
 # Copy Atom files 
 if [ ! -d "$HOME/.atom" ]; then
     mkdir $HOME/.atom
@@ -65,7 +35,18 @@ fi
 cp -a .atom/. ~/.atom/
 
 # Create Symlinks
-rm ~/.zshrc
+if [ ! -f "~/.zshrc" ]; then
+    rm ~/.zshrc
+fi
+if [ ! -f "~/.vimrc" ]; then
+    rm ~/.vimrc
+fi
 sudo ln -s $PWD/.zshrc ~/.zshrc
 sudo ln -s $PWD/.vimrc ~/.vimrc
+
+# Install Oh-My-Zsh
+sudo apt-get install zsh -y
+sh -c "$(curl -fsSL https://raw.githubusercontent.com/robbyrussell/oh-my-zsh/master/tools/install.sh)"
+sudo chsh -s $(which zsh)
+sudo apt-get install fonts-powerline -y
